+++ assets/javascripts/kitten/components/tooltips/info-with-tooltip.js
@@ -1,0 +1,4 @@
+import { withTooltip } from 'kitten/hoc/with-tooltip'
+import { ButtonInfoIcon } from 'kitten/components/buttons/button-info-icon'
+
+export const InfoWithTooltip = withTooltip(ButtonInfoIcon, {})