+++ lib/kitten/configuration.rb
@@ -149,9 +149,8 @@ module Kitten
             { path: 'components/box/information-box', name: 'Information box' },
             { path: 'components/box/link-box', name: 'Link box' },
             { path: 'components/box/link-box-with-icon', name: 'Link box with icon' },
-            { path: 'components/tooltips/tooltip', name: 'Tooltip' },
             { path: 'components/tooltips/static-tooltip', name: 'Static tooltip' },
-            { path: 'components/tooltips/new-tooltip', name: 'New tooltip' },
+            { path: 'components/tooltips/tooltip', name: 'Tooltip' },
           ],
         },
         {