+++ CHANGELOG.md
@@ -4,6 +4,12 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [unreleased]
 
+- Breaking change: Rename `ButtonTooltioIcon` to `ButtonInfoIcon`
+- Breaking change: Rename TooltipIcon` to `InfoIcon`
+- Breaking change: Split tooltip component into a component `Tooltip` and hoc:
+  `withTooltip`.
+  To be retro-compatible, you should replace `Tooltip` component with `InfoWithTooltip`.
+- Feature: Add new `NewTooltip` component.
 - Fix: Fix display element as a block element (`ButtonImageWithText`, `ButtonImage`,
        `ButtonWithBadge`, `IconBadge`).
 - Fix: Remove `React.PropTypes` and use `PropTypes` from